// src/app/interfaces/detalleRecetaSum.interface.ts
export interface IDetalleRecetaSum {
  total: number;
  ca: number;
  colesterol: number;
  fe: number;
  na: number;
  fibra: number;
  grs: number;
  h_d_c: number;
  potasio: number;
  prot: number;
  vitamina_a: number;
  vitamina_b1: number;
  vitamina_b6: number;
  vitamina_b12: number;
  vitamina_c: number;
  vitamina_e: number;
  vitamina_k: number;
}